backend/requirements.txt
flask
pandas
scikit-learn
joblib
pyjwt